binutils: make link of mixed endian objects an error
Note: The chosen method to do this seems a bit too indirect. The trap happens in
a hook for merging private sections in the object files. We end up not doing any
actual merging (and have never done any historically).

We are essentially following the methodology set by other ports. This is worth
reviewing on the mailing list at some point of time.